Fix - Gradle

+++ Task_2_2_1/src/main/java/ru/nsu/fit/oop/arturploter/task_2_2_1/Apple.java
@@ -6,7 +6,7 @@ class Apple extends GameObject {
 
   static final Color FRUIT_COLOR = Color.rgb(255, 75, 40);
 
-  Apple(int x, int y) {
+  Apple(double x, double y) {
     super(x, y);
   }
 }
