+++ Task_2_2_1/src/main/java/ru/nsu/fit/oop/arturploter/task_2_2_1/Obstacle.java
@@ -8,7 +8,7 @@ class Obstacle extends GameObject {
 
   static final int OBSTACLES_START_NUMBER = 10;
 
-  Obstacle(int x, int y) {
+  Obstacle(double x, double y) {
     super(x, y);
   }
 }
