+++ Task_2_2_1/src/main/java/ru/nsu/fit/oop/arturploter/task_2_2_1/WhiteApple.java
@@ -10,7 +10,7 @@ class WhiteApple extends GameObject {
 
   static final int SUPER_STATE_TIME = 9000;
 
-  WhiteApple(int x, int y) {
+  WhiteApple(double x, double y) {
     super(x, y);
   }
 }
